ReadCsvAndGET/pyproject.toml
[tool.poetry]
name = "readcsvandget"
version = "0.1.0"
description = ""
authors = ["Priyanka Khire <priyanka.khire@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
